+++ python/lsst/ts/envsensors/command_handler.py
@@ -25,9 +25,10 @@ import asyncio
 import logging
 import platform
 import time
-import typing
+from typing import Any, Callable, Dict, List, Optional, Union
 
 from .command_error import CommandError
+from .constants import Command, Key, DeviceType
 from .ess_instrument_object import EssInstrument
 from .mock.mock_temperature_sensor import MockTemperatureSensor
 from .response_code import ResponseCode
@@ -49,23 +50,23 @@ class CommandHandler:
 
     The commands that can be handled are:
 
-        configure: Load the _configuration that is passed on with the command
-        and connect to the devices specified in that _configuration. This
+        configure: Load the configuration that is passed on with the command
+        and connect to the devices specified in that configuration. This
         command can be sent multiple times before a start is received and only
-        the last _configuration is kept.
+        the last configuration is kept.
         start: Start reading the sensor data of the connected devices and send
-        it as plain text via the socket. If no _configuration was sent then the
-        start command is ignored. Once started no _configuration changes can be
+        it as plain text via the socket. If no configuration was sent then the
+        start command is ignored. Once started no configuration changes can be
         done anymore.
         stop: Stop sending sensor data and disconnect from all devices. Once
-        stopped, _configuration changes can be done again and/or reading of
+        stopped, configuration changes can be done again and/or reading of
         sensor data can be started again.
 
     """
 
     valid_simulation_modes = (0, 1)
 
-    def __init__(self, callback: typing.Callable, simulation_mode: int) -> None:
+    def __init__(self, callback: Callable, simulation_mode: int) -> None:
         self.log = logging.getLogger(type(self).__name__)
         if simulation_mode not in self.valid_simulation_modes:
             raise ValueError(
@@ -76,21 +77,21 @@ class CommandHandler:
         self.simulation_mode = simulation_mode
 
         self._callback = callback
-        self._configuration: typing.Optional[typing.Dict[str, typing.Any]] = None
+        self._configuration: Optional[Dict[str, Any]] = None
         self._started = False
-        self._ess_instruments: typing.List[EssInstrument] = []
+        self._ess_instruments: List[EssInstrument] = []
 
-        self.dispatch_dict = {
-            "configure": self.configure,
-            "start": self.start_sending_telemetry,
-            "stop": self.stop_sending_telemetry,
+        self.dispatch_dict: Dict[str, Callable] = {
+            Command.CONFIGURE: self.configure,
+            Command.START: self.start_sending_telemetry,
+            Command.STOP: self.stop_sending_telemetry,
         }
 
         # Unit tests may set this to an integer value to simulate a
         # disconnected or missing sensor.
         self.disconnected_channel = None
 
-    async def handle_command(self, command: str, **kwargs: typing.Any) -> None:
+    async def handle_command(self, command: str, **kwargs: Any) -> None:
         """Handle incomming commands and parameters.
 
         Parameters
@@ -104,12 +105,100 @@ class CommandHandler:
         func = self.dispatch_dict[command]
         try:
             await func(**kwargs)
-            response = {"response": ResponseCode.OK}
+            response = {Key.RESPONSE: ResponseCode.OK}
         except CommandError as e:
-            response = {"response": e.response_code}
+            response = {Key.RESPONSE: e.response_code}
         await self._callback(response)
 
-    async def configure(self, configuration: dict) -> None:
+    def _validate_configuration(self, configuration: Dict[str, Any]):
+        """Validate the configuration.
+
+        Parameters
+        ----------
+        configuration: `dict`
+            A dict representing the configuration. The format of the dict
+            follows the configuration of the ts_ess project.
+
+        Raises
+        ------
+        `CommandError`:
+            In case the provided configuration is incorrect.
+
+        """
+        # Key.DEVICES is mandatory.
+        if Key.DEVICES not in configuration:
+            raise CommandError(
+                msg=f"Missing configuration key {Key.DEVICES}.",
+                response_code=ResponseCode.INVALID_CONFIGURATION,
+            )
+        # Only one key allowed.
+        if len(configuration.keys()) != 1:
+            raise CommandError(
+                msg=f"Expected one configuration key but got {len(configuration.keys())}.",
+                response_code=ResponseCode.INVALID_CONFIGURATION,
+            )
+        # Validate the device configurations.
+        device_configurations = configuration[Key.DEVICES]
+        if not device_configurations:
+            raise CommandError(
+                msg=f"The configuration data for key {Key.DEVICES} is missing.",
+                response_code=ResponseCode.INVALID_CONFIGURATION,
+            )
+        for device_configuration in device_configurations:
+            self._validate_device_configuration(
+                device_configuration=device_configuration
+            )
+
+    def _validate_device_configuration(self, device_configuration: Dict[str, Any]):
+        """Validate the device configuration.
+
+        Parameters
+        ----------
+        device_configuration: `dict`
+            A dict representing the device configuration. The format of the
+            dict follows the configuration of the ts_ess project.
+
+        Raises
+        ------
+        `CommandError`:
+            In case the provided configuration is incorrect.
+
+        """
+        # Key.NAME, Key.CHANNELS and Key.TYPE are mandatory.
+        if (
+            Key.NAME not in device_configuration
+            or Key.CHANNELS not in device_configuration
+            or Key.TYPE not in device_configuration
+        ):
+            raise CommandError(
+                msg=f"The configuration keys {Key.NAME}, {Key.CHANNELS} and {Key.TYPE} are mandatory.",
+                response_code=ResponseCode.INVALID_CONFIGURATION,
+            )
+
+        # Make sure that Key.TYPE has the correct value.
+        if device_configuration[Key.TYPE] not in [DeviceType.FTDI, DeviceType.SERIAL]:
+            raise CommandError(
+                msg=f"The value for key {Key.TYPE} must be either {DeviceType.FTDI} or {DeviceType.SERIAL}",
+                response_code=ResponseCode.INVALID_CONFIGURATION,
+            )
+
+        # Make sure that Key.FTDI_ID is present for DeviceType.FTDI devices.
+        if device_configuration[Key.TYPE] == DeviceType.FTDI:
+            if Key.FTDI_ID not in device_configuration:
+                raise CommandError(
+                    msg=f"Missing configuration key {Key.FTDI_ID} for device of type {DeviceType.FTDI}",
+                    response_code=ResponseCode.INVALID_CONFIGURATION,
+                )
+        # Make sure that Key.SERIAL_PORT is present for DeviceType.SERIAL
+        # devices.
+        if device_configuration[Key.TYPE] == DeviceType.SERIAL:
+            if Key.SERIAL_PORT not in device_configuration:
+                raise CommandError(
+                    msg=f"Missing configuration key {Key.SERIAL_PORT} for device of type {DeviceType.SERIAL}",
+                    response_code=ResponseCode.INVALID_CONFIGURATION,
+                )
+
+    async def configure(self, configuration: Dict[str, Any]) -> None:
         """Apply the configuration.
 
         Parameters
@@ -125,13 +214,14 @@ class CommandHandler:
             ALREADY_STARTED if the command handler was started.
 
         """
-        # TODO: Implement misconfiguration handling (DM-30069)
-        self.log.info(f"configure with _configuration data {configuration}")
+        self.log.info(f"configure with configuration data {configuration}")
         if self._started:
             raise CommandError(
                 msg="Ignoring the configuration because telemetry loop already running. Send a stop first.",
                 response_code=ResponseCode.ALREADY_STARTED,
             )
+        self._validate_configuration(configuration=configuration)
+
         self._configuration = configuration
 
     async def start_sending_telemetry(self) -> None:
@@ -155,19 +245,18 @@ class CommandHandler:
 
     async def connect_devices(self) -> None:
         """Loop over the configuration and start all devices."""
-        # TODO: Implement misconfiguration handling (DM-30069)
         self.log.info("connect_devices")
-        configured_devices = self._configuration["devices"]
-        for configured_device in configured_devices:
-            device = self._get_device(configured_device)
+        device_configurations = self._configuration[Key.DEVICES]  # type: ignore
+        for device_configuration in device_configurations:
+            device = self._get_device(device_configuration)
             sel_temperature = SelTemperature(
-                name=configured_device["name"],
+                name=device_configuration[Key.NAME],
                 uart_device=device,
-                channels=configured_device["channels"],
+                channels=device_configuration[Key.CHANNELS],
                 log=self.log,
             )
             ess_instrument = EssInstrument(
-                name=configured_device["name"],
+                name=device_configuration[Key.NAME],
                 reader=sel_temperature,
                 callback_func=self._process_sensor_telemetry,
                 log=self.log,
@@ -212,17 +301,17 @@ class CommandHandler:
             The telemetry data to send.
         """
         self.log.debug(f"Processing sensor data {telemetry}")
-        data = {"telemetry": telemetry}
+        data = {Key.TELEMETRY: telemetry}
         await self._callback(data)
 
-    def _get_device(self, configured_device: dict) -> typing.Optional[typing.Any]:
-        """Get the device to connect to by using the _configuration of the CSC
+    def _get_device(self, device_configuration: dict) -> Optional[Any]:
+        """Get the device to connect to by using the configuration of the CSC
         and by detecting whether the code is running on an aarch64 architecture
         or not.
 
         Parameters
         ----------
-        configured_device: `dict`
+        device_configuration: `dict`
             A dict representing the device to connect to. The format of the
             dict follows the configuration of the ts_ess project.
 
@@ -235,34 +324,39 @@ class CommandHandler:
         Raises
         ------
         RuntimeError
-            In case an incorrect _configuration has been loaded.
+            In case an incorrect configuration has been loaded.
         """
-        device = None
+        device: Any = None
         if self.simulation_mode == 1:
             self.log.info("Connecting to the mock sensor.")
             device = MockTemperatureSensor(
-                configured_device["name"],
-                configured_device["channels"],
+                device_configuration[Key.NAME],
+                device_configuration[Key.CHANNELS],
                 disconnected_channel=self.disconnected_channel,
             )
-        elif configured_device["type"] == "FTDI":
+        elif device_configuration[Key.TYPE] == DeviceType.FTDI:
             from .vcp_ftdi import VcpFtdi
 
             device = VcpFtdi(
-                configured_device["name"], configured_device["ftdi_id"], self.log
+                device_configuration[Key.NAME],
+                device_configuration[Key.FTDI_ID],
+                self.log,
             )
-        elif self.config.type == "Serial":
+        elif device_configuration[Key.TYPE] == DeviceType.SERIAL:
             # make sure we are on a Raspberry Pi4
             if "aarch64" in platform.platform():
                 from .rpi_serial_hat import RpiSerialHat
 
                 device = RpiSerialHat(
-                    configured_device["name"], configured_device["port"], self.log
+                    device_configuration[Key.NAME],
+                    device_configuration[Key.SERIAL_PORT],
+                    self.log,
                 )
 
         if device is None:
             raise RuntimeError(
-                f"Could not get a {self.config.type!r} device on architecture "
-                f"{platform.platform()}. Please check the _configuration."
+                f"Could not get a {device_configuration['type']!r} device on "
+                f"architecture {platform.platform()}. Please check the "
+                f"configuration."
             )
         return device