+++ doc/version_history.rst
@@ -16,7 +16,9 @@ This version already includes many useful things:
 * A functioning socket server (for which the ``ts_tcpip`` socket server is used).
 * Code that reads the output of the connected sensors and sends the data via the socket server.
 * Support for USB and FTDI sensors.
+* Added support for connecting to and reading telemetry from multiple sensors.
+* Added configuration error handling.
 
 Requires:
 
-* ts_tcpip 0.1
+* ts_tcpip 0.2.0